+++ src/main/java/at/hannibal2/skyhanni/data/bazaar/BazaarSummary.kt
@@ -1,0 +1,12 @@
+package at.hannibal2.skyhanni.data.bazaar
+
+import com.google.gson.annotations.Expose
+
+data class BazaarSummary(
+    @Expose
+    val amount: Long,
+    @Expose
+    val pricePerUnit: Double,
+    @Expose
+    val orders: Long,
+)
