+++ src/main/java/at/hannibal2/skyhanni/data/bazaar/HypixelBazaarFetcher.kt
@@ -12,7 +12,6 @@ import at.hannibal2.skyhanni.utils.NEUItems
 import at.hannibal2.skyhanni.utils.SimpleTimeMark
 import at.hannibal2.skyhanni.utils.fromJson
 import com.google.gson.annotations.Expose
-import com.google.gson.annotations.SerializedName
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -53,23 +52,26 @@ object HypixelBazaarFetcher {
             }
             failedAttepmts = 0
         } catch (e: Exception) {
-            // TOOD remove
+            onError(fetchType, e)
+        }
+    }
+
+    private fun onError(fetchType: String, e: Exception) {
+        failedAttepmts++
+        if (failedAttepmts <= maxFailedAttepmts) {
+            ChatUtils.debug(
+                "Error fetching bazaar price data $fetchType from hypixel: ${e.message} " +
+                    "(failedAttepmts=$failedAttepmts)"
+            )
             e.printStackTrace()
-            failedAttepmts++
-            if (failedAttepmts <= maxFailedAttepmts) {
-                ChatUtils.debug(
-                    "Error fetching bazaar price data $fetchType from hypixel: ${e.message} " +
-                        "(failedAttepmts=$failedAttepmts)"
-                )
-                nextFetchTime = SimpleTimeMark.now() + 15.seconds
-            } else {
-                ErrorManager.logErrorWithData(
-                    e,
-                    message = "Error fetching bazaar price data from hypixel after $failedAttepmts attempts",
-                    "fetchType" to fetchType,
-                    betaOnly = true
-                )
-            }
+            nextFetchTime = SimpleTimeMark.now() + 15.seconds
+        } else {
+            ErrorManager.logErrorWithData(
+                e,
+                message = "Error fetching bazaar price data from hypixel after $failedAttepmts attempts",
+                "fetchType" to fetchType,
+                betaOnly = true
+            )
         }
     }
 
@@ -91,48 +93,4 @@ object HypixelBazaarFetcher {
         val products: Map<String, BazaarProduct>,
     )
 
-    data class BazaarProduct(
-        @Expose
-        @SerializedName("product_id")
-        val productId: String,
-        @Expose
-        @SerializedName("quick_status")
-        val quickStatus: QuickStatus,
-        @Expose
-        @SerializedName("sell_summary")
-        val sellSummary: List<Summary>,
-        @Expose
-        @SerializedName("buy_summary")
-        val buySummary: List<Summary>,
-    )
-
-    data class Summary(
-        @Expose
-        val amount: Long,
-        @Expose
-        val pricePerUnit: Double,
-        @Expose
-        val orders: Long,
-    )
-
-    class QuickStatus(
-        @Expose
-        val productId: String,
-        @Expose
-        val sellPrice: Double,
-        @Expose
-        val sellVolume: Long,
-        @Expose
-        val sellMovingWeek: Long,
-        @Expose
-        val sellOrders: Long,
-        @Expose
-        val buyPrice: Double,
-        @Expose
-        val buyVolume: Long,
-        @Expose
-        val buyMovingWeek: Long,
-        @Expose
-        val buyOrders: Long,
-    )
 }
